17-textbook-summarizer/requirements.txt
requests
rich
click
pyyaml
streamlit
pytest
pytest-cov
